requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
opencv-python>=4.5.0
scikit-image>=0.18.0
onnxruntime>=1.8.0
transformers>=4.20.0
Pillow>=8.0.0
tqdm>=4.62.0
huggingface_hub>=0.16.0 